% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addendum_stats.R
\name{addendum_stats}
\alias{addendum_stats}
\alias{addendum_stats.default}
\title{An internal method to add some text to summary statistics part}
\usage{
addendum_stats(text, addendum, type = "latex")

\method{addendum_stats}{default}(text, addendum, type = "latex")
}
\arguments{
\item{text}{Text in statistics part of the table}

\item{addendum}{Text that should be added}

\item{type}{Reporting output used. Accepted values are
\emph{latex} and \emph{html}}
}
\description{
An internal method to add some text to summary statistics part
}
