% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC.light.zeroinfl}
\alias{BIC.light.zeroinfl}
\title{Bayesian information criterion implementation for stripped objects}
\usage{
\method{BIC}{light.zeroinfl}(object, ...)
}
\arguments{
\item{object}{A fitted model object for which there exists a logLik
method to extract the corresponding log-likelihood, or
an object inheriting from class logLik.}

\item{...}{Optionally more fitted model objects

Formula \eqn{-2l + k\log(npar)} where \emph{npar}
represents the number of parameters in the fitted model
and \eqn{l} represents model log-likelihood}
}
\description{
Bayesian information criterion implementation for stripped objects
}
