% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alpha.R
\name{extract_alpha}
\alias{extract_alpha}
\alias{extract_alpha.default}
\alias{extract_alpha.zeroinfl}
\alias{extract_alpha.glm}
\title{Methods to extract dispersion parameters
in negative binomial settings}
\usage{
extract_alpha(object, ...)

\method{extract_alpha}{default}(object, ...)

\method{extract_alpha}{zeroinfl}(object, ...)

\method{extract_alpha}{glm}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\value{
A value (chracter formatted)
that can be passed to latex. Default
returns an empty string
}
\description{
Methods to extract dispersion parameters
in negative binomial settings
}
