% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coeff.R
\name{extract_coeff}
\alias{extract_coeff}
\alias{extract_coeff.default}
\alias{extract_coeff.fastLm}
\alias{extract_coeff.light.glm}
\alias{extract_coeff.oglmx}
\alias{extract_coeff.light.zeroinfl}
\alias{extract_coeff.fastzeroinfl}
\alias{extract_coeff.mindist}
\alias{extract_coeff.nnet}
\title{Methods to extract coefficients and
related information from statistical
models}
\usage{
extract_coeff(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{default}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{fastLm}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{light.glm}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{oglmx}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{light.zeroinfl}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{fastzeroinfl}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{mindist}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{extract_coeff}{nnet}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{Regression objects}

\item{...}{Additional parameters}
}
\value{
Coefficients ready to be merged
to produce a latex table
}
\description{
Methods to extract coefficients and
related information from statistical
models
}
