% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_table_coefficients.R
\name{light_table_coefficients}
\alias{light_table_coefficients}
\title{Prepare coefficients part of the latex table}
\usage{
light_table_coefficients(
  ncols_models,
  coeff_data,
  order_variable,
  omit,
  covariate.labels,
  rules_between_covariates
)
}
\arguments{
\item{order_variable}{A vector that indicates the order
in which variables will appear in the output.}

\item{omit}{List of variables that should be removed from
the table}

\item{covariate.labels}{A character vector of labels for
columns in regression tables.
Their layout, in terms of the number of columns
associated with each label, is given by the
argument \code{column.separate}.}

\item{rules_between_covariates}{A numeric vector that specifies how
rules should be laid out across rows of the regression table}
}
\description{
Prepare coefficients part of the latex table
}
