% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listcoeff.R
\name{listcoeff}
\alias{listcoeff}
\alias{listcoeff.light.zeroinfl}
\alias{listcoeff.zeroinfl}
\alias{listcoeff.light.glm}
\alias{listcoeff.default}
\alias{listcoeff.nnet}
\alias{listcoeff.light.lm}
\title{A small method to extract coefficient names from
econometric models}
\usage{
listcoeff(object, ...)

\method{listcoeff}{light.zeroinfl}(object, ...)

\method{listcoeff}{zeroinfl}(object, ...)

\method{listcoeff}{light.glm}(object, ...)

\method{listcoeff}{default}(object, ...)

\method{listcoeff}{nnet}(object, ...)

\method{listcoeff}{light.lm}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\value{
A character vector with names
that can be passed to latex
}
\description{
A small method to extract coefficient names from
econometric models
}
