% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liststats.R
\name{liststats}
\alias{liststats}
\alias{liststats.light.zeroinfl}
\alias{liststats.light.glm}
\alias{liststats.default}
\alias{liststats.light.lm}
\title{A small method to extract statistics names relevent
to econometric models}
\usage{
liststats(object, ...)

\method{liststats}{light.zeroinfl}(object, ...)

\method{liststats}{light.glm}(object, ...)

\method{liststats}{default}(object, ...)

\method{liststats}{light.lm}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments. See \code{details}}
}
\value{
A character vector with names
that can be passed to latex
}
\description{
A small method to extract statistics names relevent
to econometric models
}
\details{
The following arguments add statistics
\itemize{
\item \code{add_link = TRUE}: Add two rows informing
on the count and selection distributions
used in \code{object} (possibly empty)
\item \code{add_alpha = TRUE}: Add a row on the
dispersion parameter for negative binomial
models (possibly empty)
\item \code{add_sigma = TRUE}: Add a row on the
estimated standard deviation (see \link[stats]{sigma})
}
}
