% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_light.R
\name{logLik.light.zeroinfl}
\alias{logLik.light.zeroinfl}
\alias{logLik.light.glm}
\title{Extract loglikelihood from a stripped object}
\usage{
\method{logLik}{light.zeroinfl}(object, ...)

\method{logLik}{light.glm}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\value{
Returs loglikelihood
}
\description{
Extends \link[stats]{logLik} to objects of
class \code{light}* (i.e. regression objects
after applying \link{strip} method)
}
\details{
Implementation, at this stage,
only returns the loglikelihood value.
Additional returns might be included
in future releases
}
\seealso{
\link[stats]{logLik}
}
