% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multiple gsub}
\usage{
mgsub(pattern, replacement, text.var, fixed = TRUE, ...)
}
\arguments{
\item{pattern}{Character string to be matched in the given character vector.}

\item{replacement}{Character string equal in length to pattern or of length
one which are  a replacement for matched pattern.}

\item{text.var}{The text variable.}

\item{fixed}{logical. If \code{TRUE}, pattern is a string to be matched as is.
Overrides all conflicting arguments.}

\item{\dots}{Additional arguments passed to \code{\link[base]{gsub}}.}

\item{leadspace}{logical.  If \code{TRUE} inserts a leading space in the
replacements.}

\item{trailspace}{logical.  If \code{TRUE} inserts a trailing space in the
replacements.}

\item{trim}{logical.  If \code{TRUE} leading and trailing white spaces are
removed and multiple white spaces are reduced to a single white space.}

\item{order.pattern}{logical.  If \code{TRUE} and \code{fixed = TRUE}, the
\code{pattern} string is sorted by number of characters to prevent substrings
replacing meta strings (e.g., \code{pattern = c("the", "then")} resorts to
search for "then" first).}
}
\value{
\code{multigsub} - Returns a vector with the pattern replaced.
}
\description{
\code{multigsub} - Borrowed from \code{qdap} library.
A wrapper for \code{\link[base]{gsub}} that takes a vector
of search terms and a vector or single value of replacements.
}
\examples{
\dontrun{
## ======================
##    `mgsub` Function
## ======================

mgsub(c("it's", "I'm"), c("it is", "I am"), DATA$state)
mgsub("[[:punct:]]", "PUNC", DATA$state, fixed = FALSE)

}
}
\seealso{
\code{\link[base]{gsub}}
}
