% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_nnet_reference.R
\name{reorder_nnet_reference}
\alias{reorder_nnet_reference}
\alias{reorder_nnet_reference.nnet}
\alias{reorder_nnet_reference.list}
\title{Reorder columns for nnet models}
\usage{
reorder_nnet_reference(
  object,
  coeff_body,
  reference_level_position = NULL,
  type,
  ...
)

\method{reorder_nnet_reference}{nnet}(
  object,
  coeff_body,
  reference_level_position = NULL,
  type,
  ...
)

\method{reorder_nnet_reference}{list}(
  object,
  coeff_body,
  reference_level_position = NULL,
  type,
  ...
)
}
\arguments{
\item{object}{List of object}

\item{coeff_body}{Coefficient body part of the regression table}

\item{reference_level_position}{In multinomial logit, coefficients are
relative to the reference level. A column can be added to represent
an empty coefficient modality that is indicated as being the reference.
If \code{NULL}, this feature is ignored. Otherwise, the column is placed at
the position indicated by the argument}

\item{type}{Reporting output used. Accepted values are
\emph{latex} and \emph{html}}

\item{...}{Additional parameters, currently ignored}
}
\value{
Initial object (latex or html code for coefficients
reporting) with the reference level at desired column
}
\description{
Reorder columns for nnet models
}
