% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_stars.R
\name{signif_stars}
\alias{signif_stars}
\title{Transform pvalues into significance stars}
\usage{
signif_stars(pvalue)
}
\arguments{
\item{pvalue}{pvalue for coefficients}
}
\description{
Transform pvalues into significance stars
}
