% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_regex.R
\name{str_to_regex}
\alias{str_to_regex}
\title{Espace special characters in regex}
\usage{
str_to_regex(str)
}
\arguments{
\item{str}{A character vector}
}
\value{
Initial vector with special
characters escaped
}
\description{
Espace special characters in regex
}
