% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip}
\alias{strip}
\alias{strip.glm}
\alias{strip.lm}
\alias{strip.zeroinfl}
\alias{strip.summary.glm}
\alias{strip.selection}
\alias{strip.summary.lm}
\alias{strip.summary.selection}
\alias{strip.summary.zeroinfl}
\title{strip: strip regressions}
\usage{
strip(object, ...)

\method{strip}{glm}(object)

\method{strip}{lm}(object)

\method{strip}{zeroinfl}(object)

\method{strip}{summary.glm}(object)

\method{strip}{selection}(object)

\method{strip}{summary.lm}(object)

\method{strip}{summary.selection}(object)

\method{strip}{summary.zeroinfl}(object)
}
\arguments{
\item{object}{A regression object or summary.
Accepted classes are \link[stats]{glm} or
\link[pscl]{zeroinfl}}

\item{...}{Additional arguments that should be
considered}
}
\value{
Initial object stripped from
heavy elements. Initial class is returned
with a new one indicating the object has
been stripped.
}
\description{
Strip regression objects to reduce memory needs
}
\details{
Remove elements that are more related
to prediction than model to reduce
memory needs to store a glm output
}
