% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablelight.R
\docType{package}
\name{tablelight}
\alias{tablelight}
\title{lighttable: Regression tables using as little as possible RAM resources}
\description{
This package provides with two types of functions: some that
strip regression objects from unnecessary elements to
produce tables (\code{strip} family) and others that produce
tables (\code{light_table})
}
\section{strip}{

blabla
}

\section{light_table}{

blabla
}

