% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{variance_model}
\alias{variance_model}
\title{Construct the dataframe used for
variance estimation}
\usage{
variance_model(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \link[oglmx]{oglmx} object}

\item{newdata}{Dataframe that must be used. An
error will be thrown away if the model matrix
for variance cannot be constructed}

\item{...}{Additional arguments}
}
\description{
Construct the dataframe used for
variance estimation
}
