% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.R
\name{view_html}
\alias{view_html}
\title{Report HTML table in Rstudio viewer}
\usage{
view_html(x)
}
\description{
@param x HTML code
@return Returns nothing, HTML reported in
Rstudio viewer
@importFrom rstudioapi viewer
}
