% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_table_coefficients.R
\name{light_table_coefficients}
\alias{light_table_coefficients}
\title{Prepare coefficients part of the latex table}
\usage{
light_table_coefficients(
  object,
  ncols_models,
  type,
  coeff_data,
  order_variable,
  omit,
  covariate.labels,
  reference_level_position,
  rules_between_covariates
)
}
\arguments{
\item{object}{List of object}

\item{ncols_models}{Number of columns}

\item{type}{Reporting output used. Accepted values are
\emph{latex} and \emph{html}}

\item{coeff_data}{Output from \link{secoeff}}

\item{order_variable}{A vector that indicates the order
in which variables will appear in the output.}

\item{omit}{List of variables that should be removed from
the table}

\item{covariate.labels}{A character vector of labels for
columns in regression tables.
Their layout, in terms of the number of columns
associated with each label, is given by the
argument \code{column.separate}.}

\item{reference_level_position}{In multinomial logit, coefficients are
relative to the reference level. A column can be added to represent
an empty coefficient modality that is indicated as being the reference.
If \code{NULL}, this feature is ignored. Otherwise, the column is placed at
the position indicated by the argument

This function is designed to produce \code{latex} tables with
stripped objects (see \link{strip}). It follows
\code{stargazer} standards but proposes a
simplified framework. Customization is limited}

\item{rules_between_covariates}{A numeric vector that specifies how
rules should be laid out across rows of the regression table}
}
\description{
Prepare coefficients part of the latex table
}
