% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liststats.R
\name{liststats}
\alias{liststats}
\alias{liststats.light.zeroinfl}
\alias{liststats.light.glm}
\alias{liststats.default}
\title{A small method to extract statistics names relevent
to econometric models}
\usage{
liststats(object, ...)

\method{liststats}{light.zeroinfl}(object, ...)

\method{liststats}{light.glm}(object, ...)

\method{liststats}{default}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\value{
A character vector with names
that can be passed to latex
}
\description{
A small method to extract statistics names relevent
to econometric models
}
