% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncolsmodels.R
\name{ncolsmodels}
\alias{ncolsmodels}
\alias{ncolsmodels.default}
\alias{ncolsmodels.nnet}
\alias{ncolsmodels.list}
\title{A method to infer the number of columns for the regression table}
\usage{
ncolsmodels(object)

\method{ncolsmodels}{default}(object)

\method{ncolsmodels}{nnet}(object)

\method{ncolsmodels}{list}(object)
}
\arguments{
\item{object}{Regression object or list of regression objects}
}
\value{
Number of columns necessary that will be used
for the regression table
}
\description{
A method to infer the number of columns for the regression table
}
