% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_coeff.R
\name{secoeff}
\alias{secoeff}
\alias{secoeff.default}
\alias{secoeff.oglmx}
\alias{secoeff.light.zeroinfl}
\alias{secoeff.negbin}
\alias{secoeff.summary.lm}
\alias{secoeff.summary.glm}
\title{Extract summary table for coefficients
estimates}
\usage{
secoeff(object, ...)

\method{secoeff}{default}(object, ...)

\method{secoeff}{oglmx}(object, ...)

\method{secoeff}{light.zeroinfl}(object, ...)

\method{secoeff}{negbin}(object, ...)

\method{secoeff}{summary.lm}(object, ...)

\method{secoeff}{summary.glm}(object, ...)
}
\arguments{
\item{object}{A regression object}

\item{...}{Additional parameters}
}
\value{
Returns coefficient values,
standard errors, t-stat (or z-stat)
and pvalues
}
\description{
Extract summary table for coefficients
estimates
}
