% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip}
\alias{strip}
\alias{strip.glm}
\alias{strip.fastglm}
\alias{strip.oglmx}
\alias{strip.negbin}
\alias{strip.lm}
\alias{strip.zeroinfl}
\alias{strip.fastnegbin}
\alias{strip.fastzeroinfl}
\alias{strip.summary.glm}
\alias{strip.summary.lm}
\alias{strip.summary.zeroinfl}
\title{strip: strip regressions}
\usage{
strip(object, ...)

\method{strip}{glm}(object, ...)

\method{strip}{fastglm}(object, ...)

\method{strip}{oglmx}(object, ...)

\method{strip}{negbin}(object, ...)

\method{strip}{lm}(object, ...)

\method{strip}{zeroinfl}(object, ...)

\method{strip}{fastnegbin}(object, ...)

\method{strip}{fastzeroinfl}(object, ...)

\method{strip}{summary.glm}(object, ...)

\method{strip}{summary.lm}(object, ...)

\method{strip}{summary.zeroinfl}(object, ...)
}
\arguments{
\item{object}{A regression object or summary.
Accepted classes are \link[stats]{glm} or
\link[pscl]{zeroinfl}}

\item{...}{Additional arguments that should be
considered}
}
\value{
Initial object stripped from
heavy elements. Initial class is returned
with a new one indicating the object has
been stripped.
}
\description{
Strip regression objects to reduce memory needs
}
\details{
Remove elements that are more related
to prediction than model to reduce
memory needs to store a glm output
}
