% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.R
\name{view_html}
\alias{view_html}
\title{Report HTML table in Rstudio viewer}
\usage{
view_html(x)
}
\arguments{
\item{x}{HTML code}
}
\value{
Returns nothing, HTML reported in
Rstudio viewer
}
\description{
Report HTML table in Rstudio viewer
}
