% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.R
\name{view_html}
\alias{view_html}
\alias{print_html}
\title{Report HTML table in Rstudio viewer or in a Rmarkdown document}
\usage{
view_html(x)

print_html(x)
}
\arguments{
\item{x}{HTML code}
}
\value{
Returns nothing, HTML reported in
Rstudio viewer (\code{view_html}) or in
the Rmd document (\code{print_html})
}
\description{
Report HTML table in Rstudio viewer or in a Rmarkdown document
}
